pyproject.toml
[tool.poetry]
name = "websitehendrybackend"
version = "0.1.0"
description = "Backend for my personal website (www.hendrywidyanto.com)"
authors = ["Hendry Widyanto <hendrywidyanto97@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11.3"
azure-functions = "^1.18.0"
fastapi = "^0.110.1"
pandas = "^2.2.2"
python-dotenv = "^1.0.1"
SQLAlchemy = "^2.0.29"
pyodbc = "^5.1.0"
alembic = "^1.13.1"
pydantic = "^2.6.4"
uvicorn = {extras = ["standard"], version = "^0.29.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
